requirements.txt
flask
flask-cors
tensorflow-cpu
opencv-python
numpy
